move strided_memcpy
