+++ paddle/phi/tests/kernels/CMakeLists.txt
@@ -91,3 +91,8 @@ cc_test(
   test_cache
   SRCS test_cache.cc
   DEPS gtest cache)
+
+cc_test(
+  strided_memcpy_test
+  SRCS strided_memcpy_test.cc
+  DEPS tensor memory)
